Optimize server observers registration phase by skipping runOnLocks(nick)
Uses the badge pattern to allow the ServerController (and only him)
to call a version of PlayerObservableTracker#registerObserver() which
already assumes locking, therefore skipping the lock search.
This basically brings it down from O(k*n) to O(k), where k is the number of
observers to register (so a constant factor) and n the number of lobbies.