Change junit timeout thread mode to SEPARATE_THREAD
SAME_THREAD mode (the default) only calls the interrupt method
on the thread executing the test after the timeout expires.
While this should work, in practice there are cases where it's
simply ignored, so let's use SEPARATE_THREAD, which will actually
fail the test after the timeout regardless of what the testing
code does.

Since we don't use ThreadLocal storage, we should not be impacted
by the thread-offs mentioned in the junit docs for preemtpive
timeouts.